Be more careful `updateRepositoryUpdate()` handling the `build-extra` updates
The `build-extra` updates commit some changes before pushing. Since we
want to use this code also for embargoed releases, we need to decouple
this so that a "phase 1" workflow run can commit these changes, and a
"phase 2" workflow run can push the changes to the non-embargoed
repository.

Technically, this is still not quite correct: The last part, the one
updating `download-stats.sh`, can only be done once the embargo has been
lifted (because it required the ID of the public GitHub Release, which
will only be available at that time).

+++ repository-updates.js
@@ -100,31 +100,35 @@ const pushRepositoryUpdate = async (context, setSecret, appId, privateKey, owner
   if (repo === 'build-extra') {
     // Add `versions/package-versions-$ver*.txt`
     const fs = require('fs')
-    const ver = fs.readFileSync('bundle-artifacts/ver').toString().trim()
-    fs.renameSync(
-      'installer-x86_64/package-versions.txt',
-      `${repo}/versions/package-versions-${ver}.txt`
-    )
-    fs.renameSync(
-      'mingit-x86_64/package-versions.txt',
-      `${repo}/versions/package-versions-${ver}-MinGit.txt`
-    )
-    callGit([
-      'add',
-      `versions/package-versions-${ver}.txt`,
-      `versions/package-versions-${ver}-MinGit.txt`
-    ], repo)
-    callGit([
-      'commit',
-      '-s',
-      '-m', `versions: add v${ver}`,
-      `versions/package-versions-${ver}.txt`,
-      `versions/package-versions-${ver}-MinGit.txt`
-    ], repo)
+    if (fs.existsSync('bundle-artifacts/ver')) {
+      const ver = fs.readFileSync('bundle-artifacts/ver').toString().trim()
+      fs.renameSync(
+        'installer-x86_64/package-versions.txt',
+        `${repo}/versions/package-versions-${ver}.txt`
+      )
+      fs.renameSync(
+        'mingit-x86_64/package-versions.txt',
+        `${repo}/versions/package-versions-${ver}-MinGit.txt`
+      )
+      callGit([
+        'add',
+        `versions/package-versions-${ver}.txt`,
+        `versions/package-versions-${ver}-MinGit.txt`
+      ], repo)
+      callGit([
+        'commit',
+        '-s',
+        '-m', `versions: add v${ver}`,
+        `versions/package-versions-${ver}.txt`,
+        `versions/package-versions-${ver}-MinGit.txt`
+      ], repo)
+    }
 
-    // Update `download-stats.sh`
-    callProg('sh', ['./download-stats.sh', '--update'], repo)
-    callGit(['commit', '-s', '-m', 'download-stats: new Git for Windows version', './download-stats.sh'], repo)
+    if (owner === 'git-for-windows' && refName === 'main') {
+      // Update `download-stats.sh`
+      callProg('sh', ['./download-stats.sh', '--update'], repo)
+      callGit(['commit', '-s', '-m', 'download-stats: new Git for Windows version', './download-stats.sh'], repo)
+    }
   } else if (repo === 'git-for-windows.github.io') {
     callGit(['switch', '-C', 'main', 'origin/main'], repo)
     callProg('node', ['bump-version.js', '--auto'], repo)
